chore: split query namespace classes
