+++ src/Query/Builders/Common/BuilderBase.cs
