+++ src/Query/Builders/Common/BuilderValidation.cs
