+++ src/Query/Builders/Common/JoinLimitationEnforcer.cs
