+++ src/Query/Builders/Functions/KsqlFunctionRegistry.cs
