+++ src/Query/Builders/Functions/KsqlFunctionTranslator.cs
