+++ src/Query/Linq/IJoinResult.cs
@@ -1,0 +1,20 @@
+using Kafka.Ksql.Linq.Core.Abstractions;
+using Kafka.Ksql.Linq.Query.Abstractions;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Kafka.Ksql.Linq.Query.Linq;
+
+public interface IJoinResult<TOuter, TInner>
+    where TOuter : class
+    where TInner : class
+{
+    IEntitySet<TResult> Select<TResult>(Expression<Func<TOuter, TInner, TResult>> resultSelector) where TResult : class;
+    IJoinResult<TOuter, TInner, TThird> Join<TThird, TKey>(IEntitySet<TThird> third, Expression<Func<TOuter, TKey>> outerKeySelector, Expression<Func<TThird, TKey>> thirdKeySelector) where TThird : class;
+    IJoinResult<TOuter, TInner, TThird> Join<TThird, TKey>(IEntitySet<TThird> third, Expression<Func<TInner, TKey>> innerKeySelector, Expression<Func<TThird, TKey>> thirdKeySelector) where TThird : class;
+}
+
