+++ src/Query/Pipeline/DDLQueryGenerator.cs
@@ -380,33 +380,3 @@ internal class DDLQueryGenerator : GeneratorBase, IDDLQueryGenerator
         };
     }
 }
-
-/// <summary>
-/// メソッド呼び出し収集Visitor
-/// </summary>
-internal class MethodCallCollectorVisitor : ExpressionVisitor
-{
-    public List<MethodCallExpression> MethodCalls { get; } = new();
-
-    protected override Expression VisitMethodCall(MethodCallExpression node)
-    {
-        MethodCalls.Add(node);
-        return base.VisitMethodCall(node);
-    }
-}
-
-/// <summary>
-/// 式解析結果（簡略版）
-/// </summary>
-internal class ExpressionAnalysisResult
-{
-    public List<MethodCallExpression> MethodCalls { get; set; } = new();
-    public bool HasGroupBy => MethodCalls.Any(mc => mc.Method.Name == "GroupBy");
-    public bool HasAggregation => MethodCalls.Any(mc => IsAggregateMethod(mc.Method.Name));
-    public bool HasWindow => MethodCalls.Any(mc => mc.Method.Name == "Window");
-
-    private static bool IsAggregateMethod(string methodName)
-    {
-        return methodName is "Sum" or "Count" or "Max" or "Min" or "Average" or "Aggregate";
-    }
-}
