+++ src/Query/Pipeline/GeneratorBase.cs
