+++ src/Query/Pipeline/JoinQueryGenerator.cs
